cart system added

+++ src/modules/category/category.service.ts
@@ -1,3 +1,4 @@
+import { AppError } from "../../errors/AppError";
 import { prisma } from "../../lib/prisma";
 
 const createCategory = async (payload: { name: string }) => {
@@ -12,7 +13,21 @@ const getAllCategories = async () => {
   });
 };
 
+// const updateCategory = async (id: string, name: string) => {
+//   return prisma.category.update({
+//     where: { id },
+//     data: { name },
+//   });
+// };
 const updateCategory = async (id: string, name: string) => {
+  const existing = await prisma.category.findUnique({
+    where: { id },
+  });
+
+  if (!existing) {
+    throw new AppError(404, "Category not found");
+  }
+
   return prisma.category.update({
     where: { id },
     data: { name },
@@ -20,6 +35,14 @@ const updateCategory = async (id: string, name: string) => {
 };
 
 const deleteCategory = async (id: string) => {
+  const existing = await prisma.category.findUnique({
+    where: { id },
+  });
+
+  if (!existing) {
+    throw new AppError(404, "Category not found");
+  }
+
   return prisma.category.delete({
     where: { id },
   });
